+++ backend/src/main/java/momo/app/auth/jwt/filter/JwtAuthenticationProcessingFilter.java
@@ -7,7 +7,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import momo.app.auth.dto.AuthUser;
-import momo.app.auth.jwt.service.JwtService;
+import momo.app.auth.jwt.service.JwtCreateAndUpdateService;
+import momo.app.auth.jwt.service.JwtExtractService;
+import momo.app.auth.jwt.service.JwtSendService;
 import momo.app.user.domain.User;
 import momo.app.user.domain.UserRepository;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -34,7 +36,9 @@ public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
 
     private static final String NO_CHECK_URL = "/login"; // '/login'으로 들어오는 요청은 Filter사용 X
 
-    private final JwtService jwtService;
+    private final JwtCreateAndUpdateService jwtCreateAndUpdateService;
+    private final JwtExtractService jwtExtractService;
+    private final JwtSendService jwtSendService;
     private final UserRepository userRepository;
 
     private GrantedAuthoritiesMapper authoritiesMapper = new NullAuthoritiesMapper();
@@ -46,8 +50,8 @@ public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
             return;
         }
 
-        String refreshToken = jwtService.extractRefreshToken(request) //사용자 요청 header에서 refresh 토큰 추출
-                .filter(jwtService::isTokenValid) //토큰 유효성 검사
+        String refreshToken = jwtExtractService.extractRefreshToken(request) //사용자 요청 header에서 refresh 토큰 추출
+                .filter(jwtCreateAndUpdateService::isTokenValid) //토큰 유효성 검사
                 .orElse(null); //토큰이 없거나 유효하지 않으면 null 반환
 
         // refresh토큰이 사용자 header에 존재, access 토큰이 없어서 refresh 토큰을 요청한 경우
@@ -67,24 +71,30 @@ public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
         userRepository.findByRefreshToken(refreshToken) //refresh 토큰으로 user 찾음
                 .ifPresent(user -> { // user가 존재하면
                     String reIssueRefreshToken = reIssueRefreshToken(user); // refresh 토큰 재발급, DB 업데이트
-                    jwtService.sendAccessAndRefreshToken(response, jwtService.createAccessToken(user.getEmail()), reIssueRefreshToken);
+                    jwtSendService.sendAccessAndRefreshToken(response, jwtCreateAndUpdateService.createAccessToken(user.getEmail()), reIssueRefreshToken);
                     //header에 access 토큰과 refresh 토큰 담아 보냄
                 });
     }
 
     //refresh 토큰 재발급 후 refresh 토큰 DB에 업데이트
     private String reIssueRefreshToken(User user) {
-        String reIssuedRefreshToken = jwtService.createRefreshToken();
-        user.updateRefreshToken(reIssuedRefreshToken);
-        userRepository.saveAndFlush(user);
+        String reIssuedRefreshToken = jwtCreateAndUpdateService.createRefreshToken();
+        updateUserRefreshToken(user, reIssuedRefreshToken);
         return reIssuedRefreshToken;
     }
 
+    //user의 refresh 토큰 업데이트 후 저장
+    private void updateUserRefreshToken(User user, String reIssuedRefreshToken) {
+        user.updateRefreshToken(reIssuedRefreshToken);
+        userRepository.saveAndFlush(user);
+    }
+
+
     // access 토큰 유효성 검사 및 인증 처리
     private void checkAccessTokenAndAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)  throws ServletException, IOException{
-        jwtService.extractAccessToken(request) //사용자 요청 header에서 access 토큰 추출
-                .filter(jwtService::isTokenValid) // 토큰 유효성 검사
-                .ifPresent(accessToken -> jwtService.extractEmail(accessToken) // 유효한 토큰이 있으면 이메일 추출
+        jwtExtractService.extractAccessToken(request) //사용자 요청 header에서 access 토큰 추출
+                .filter(jwtCreateAndUpdateService::isTokenValid) // 토큰 유효성 검사
+                .ifPresent(accessToken -> jwtExtractService.extractEmail(accessToken) // 유효한 토큰이 있으면 이메일 추출
                         .ifPresent(email -> userRepository.findByEmail(email) // 추출된 이메일이 존재하면 이메일로 user 검색
                                 .ifPresent(this::saveAuthentication))); // 인증 허가 메소드 실행
         filterChain.doFilter(request, response); //다음 필터로 넘어가도록 설정
@@ -93,16 +103,11 @@ public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
     //Security를 사용하여 사용자를 인증 및 허가
     private void saveAuthentication(User user) {
         AuthUser authUser = AuthUser.createAuthUser(user);
-        /*
-        UserDetails userDetails = org.springframework.security.core.userdetails.User.builder()
-                .username(user.getEmail())
-                .password(password)
-                .roles(user.getRole().name())
-                .build();
-        */
+
         //UserDetails 객체를 사용하여 사용자의 인증 정보를 나타내는 토큰 생성
         Authentication authentication = new UsernamePasswordAuthenticationToken(
                 authUser, null, authoritiesMapper.mapAuthorities(authUser.getAuthorities()));
         SecurityContextHolder.getContext().setAuthentication(authentication); //SecurityContextHolder에 생성된 인증 객체를 설정하여 사용자의 인증 정보를 저장
     }
+
 }
